feat: create helper to compare arrays of Objects
